Making InFilter the default and deprecating OrFilter of SelectorFilters
