+++ fili-core/src/main/resources/moduleConfig.properties
@@ -168,7 +168,7 @@ bard__log_formatter_implementation=com.yahoo.bard.webservice.logging.JsonLogForm
 bard__druid_registered_lookup_tiers=__default
 
 # If this flag is set to true, InFilter will be used as the default filter. Else OrFilter will be used as the default filter.
-bard__default_in_filter_enabled=false
+bard__default_in_filter_enabled=true
 
 # Truncate max filename length before suffix
 # -1 indicates no truncation, 215 currently supports limits of Excel in MacOS
