+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/endpoints/SingleDimensionNoRowsFilterDataServletSpec.groovy
@@ -77,24 +77,14 @@ class SingleDimensionNoRowsFilterDataServletSpec extends BaseDataServletComponen
             "filter": {
                 "fields": [
                     {
-                        "fields": [
-                            {
-                                "dimension": "misc",
-                                "type": "selector",
-                                "value": "other1"
-                            }
-                        ],
-                        "type": "or"
+                        "dimension": "misc",
+                        "type": "in",
+                        "values": [ "other1" ]
                     },
                     {
-                        "fields": [
-                            {
-                                "dimension": "misc",
-                                "type": "selector",
-                                "value": "other2Desc"
-                            }
-                        ],
-                        "type": "or"
+                        "dimension": "misc",
+                        "type": "in",
+                        "values": [ "other2Desc" ]
                     }
                 ],
                 "type": "and"
