+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/endpoints/SingleDimensionSimpleFilterDataServletSpec.groovy
@@ -64,14 +64,9 @@ class SingleDimensionSimpleFilterDataServletSpec extends BaseDataServletComponen
                 "model"
             ],
             "filter": {
-                "fields": [
-                    {
-                        "dimension": "model",
-                        "type": "selector",
-                        "value": "Model1"
-                    }
-                ],
-                "type": "or"
+                "dimension": "model",
+                "type": "in",
+                "values": [ "Model1" ]
             },
             "aggregations": [
                 { "name": "width", "fieldName": "width", "type": "longSum" }
