+++ fili-sql/src/test/groovy/com/yahoo/bard/webservice/sql/PrestoSqlBackedClientSpec.groovy
@@ -1,5 +1,7 @@
 // Copyright 2019 Yahoo Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
+package com.yahoo.bard.webservice.sql
+
 import com.yahoo.bard.webservice.sql.presto.PrestoDataSource
 import com.yahoo.bard.webservice.sql.presto.PrestoSqlBackedClient
 import com.yahoo.bard.webservice.table.resolver.NoMatchFoundException
