+++ luthier/src/test/resources/testApplicationConfig.properties
@@ -87,7 +87,7 @@ bard__intersection_reporting_enabled = false
 bard__case_sensitive_keys_enabled = false
 
 # If this flag is set to true, InFilter will be used as the default filter. Else OrFilter will be used as the default filter.
-bard__default_in_filter_enabled = false
+bard__default_in_filter_enabled = true
 
 # Whether or not to require metrics in queries. Older druid versions require
 bard__require_metrics_in_query = true
