+++ sample-applications/luthier-sample-app/src/main/resources/applicationConfig.properties
@@ -70,7 +70,7 @@ bard__top_n_enabled = true
 bard__case_sensitive_keys_enabled = false
 
 # If this flag is set to true, InFilter will be used as the default filter. Else OrFilter will be used as the default filter.
-bard__default_in_filter_enabled = false
+bard__default_in_filter_enabled = true
 
 # Whether or not to require metrics in queries. Older druid versions require
 bard__require_metrics_in_query = true
